task 2 & task 3 corrected

+++ src/main/java/com/kodilla/hibernate/invoice/dao/InvoiceDao.java
@@ -7,5 +7,5 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
 @Repository
-public interface InvoiceDao extends CrudRepository<Invoice, Integer> {
+public interface InvoiceDao extends CrudRepository<Invoice, Long> {
 }
