+++ src/main/java/com/kodilla/hibernate/invoice/dao/ItemDao.java
@@ -7,5 +7,5 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
 @Repository
-public interface ItemDao extends CrudRepository<Item, Integer> {
+public interface ItemDao extends CrudRepository<Item, Long> {
 }
